+++ api/src/middlewares/errorHandler.ts
@@ -5,6 +5,7 @@ import {
   NotAuthenticatedError,
   NotFoundError,
 } from "@gram/core/dist/util/errors.js";
+import { ZodError } from "zod";
 
 const log = log4js.getLogger("app");
 
@@ -18,8 +19,17 @@ export default function errorHandler(err: any, req: any, res: any, next: any) {
   } else if (err instanceof NotFoundError) {
     res.status(404);
     log.info(err); // Happens on model not found etc, not necessarily an error.
-  } else if (err instanceof InvalidInputError) {
+  } else if (err instanceof InvalidInputError || err instanceof ZodError) {
     res.status(400);
+    if (err instanceof ZodError) {
+      res.json({
+        error: err.issues,
+      });
+    } else {
+      res.json({
+        error: err.message,
+      });
+    }
     log.error(err, { errorHandled: true });
   } else {
     res.status(500);