+++ core/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.9.0](https://github.com/klarna/gram/compare/v4.8.1...v4.9.0) (2024-01-29)
+
+### Bug Fixes
+
+- control/ops api not correctly routing healthchecks and metadata. Also fix healthchecks with faulty logic. Adds new healthcheck for faulty action item exports. ([9d01102](https://github.com/klarna/gram/commit/9d01102e028fb4936cb71712f8b678773e1bf899))
+
+### Features
+
+- Ability to add custom links to threats/controls ([b237532](https://github.com/klarna/gram/commit/b237532dbdbfffc2c476fc0f6c45d332a9fcb817))
+- add new ActionItemExporter functionality ([dc5f6d5](https://github.com/klarna/gram/commit/dc5f6d5fa9439b6c354f0dc8602086f7722e13da)), closes [#61](https://github.com/klarna/gram/issues/61)
+- add proxying to jira plugin ([fe87616](https://github.com/klarna/gram/commit/fe8761653a718d0ed3d2004bd2435f5963b03ea1))
+- add the ability to export action items outside of the review flow. Also make the feature to automatically exporter action items on review approve a boolean config option. ([9943fff](https://github.com/klarna/gram/commit/9943fff9f23876e38aa86ff072a76f3b4243d2d5))
+- exported action items are also copied on imported models ([6f549e8](https://github.com/klarna/gram/commit/6f549e8a639f10a6dc103c84fb1b8128c73468f4))
+
 ## [4.8.1](https://github.com/klarna/gram/compare/v4.8.0...v4.8.1) (2024-01-02)
 
 **Note:** Version bump only for package @gram/core